Split up combat.kts
