+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/combat/CombatBlock.kts
@@ -1,0 +1,15 @@
+import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.npc.NPC
+import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
+import world.gregs.voidps.engine.event.on
+import world.gregs.voidps.world.interact.entity.combat.CombatHit
+import world.gregs.voidps.world.interact.entity.sound.playSound
+
+on<CombatHit> { character: Character ->
+    val name = (character as? NPC)?.def?.getOrNull("category") ?: "player"
+    if (source is Player) {
+        source.playSound("${name}_hit", delay = 40)
+    }
+    character.setAnimation("${name}_hit")
+}
